tools/git_scripts: git-find-order: Add next trees support
As known, the commits lying in the next trees (e.g. net-next, rdma-for-next)
which are intended for the next upstream kernel release are tagless,
therefore the program crashed when the upstream tree set to be
rdma-for-next (while trying to query a certain commit which are not
merged into linus's tree yet for a tag).

Also, remove all the debug prints from the code.

Signed-off-by: Erez Alfasi <ereza@mellanox.com>